fix(storage): Fix SocketTimeoutException when executing a long multi-part upload
