+++ aws-storage-s3/src/main/java/com/amplifyframework/storage/s3/transfer/worker/BaseTransferWorker.kt
@@ -15,20 +15,10 @@
 
 package com.amplifyframework.storage.s3.transfer.worker
 
-import android.app.NotificationChannel
-import android.app.NotificationManager
 import android.content.Context
 import android.net.ConnectivityManager
 import android.net.NetworkCapabilities
 import android.os.Build
-import android.util.Log
-import androidx.annotation.RequiresApi
-import androidx.core.app.NotificationCompat
-import androidx.work.CoroutineWorker
-import androidx.work.Data
-import androidx.work.ForegroundInfo
-import androidx.work.WorkerParameters
-import androidx.work.workDataOf
 import aws.sdk.kotlin.services.s3.model.ObjectCannedAcl
 import aws.sdk.kotlin.services.s3.model.PutObjectRequest
 import aws.sdk.kotlin.services.s3.model.RequestPayer
@@ -37,40 +27,15 @@ import aws.sdk.kotlin.services.s3.model.StorageClass
 import aws.smithy.kotlin.runtime.content.ByteStream
 import aws.smithy.kotlin.runtime.content.fromFile
 import aws.smithy.kotlin.runtime.time.Instant
-import com.amplifyframework.core.Amplify
-import com.amplifyframework.core.category.CategoryType
 import com.amplifyframework.storage.ObjectMetadata
-import com.amplifyframework.storage.TransferState
-import com.amplifyframework.storage.s3.AWSS3StoragePlugin
-import com.amplifyframework.storage.s3.R
 import com.amplifyframework.storage.s3.transfer.ProgressListener
-import com.amplifyframework.storage.s3.transfer.TransferDB
 import com.amplifyframework.storage.s3.transfer.TransferRecord
-import com.amplifyframework.storage.s3.transfer.TransferStatusUpdater
 import java.io.File
-import java.lang.Exception
-import java.net.SocketException
-import kotlinx.coroutines.CancellationException
-import kotlinx.coroutines.currentCoroutineContext
-import kotlinx.coroutines.isActive
 
 /**
  * Base worker to perform transfer file task.
  */
-internal abstract class BaseTransferWorker(
-    private val transferStatusUpdater: TransferStatusUpdater,
-    private val transferDB: TransferDB,
-    context: Context,
-    workerParameters: WorkerParameters
-) : CoroutineWorker(context, workerParameters) {
-
-    internal lateinit var transferRecord: TransferRecord
-    internal lateinit var outputData: Data
-    private val logger =
-        Amplify.Logging.logger(
-            CategoryType.STORAGE,
-            AWSS3StoragePlugin.AWS_S3_STORAGE_LOG_NAMESPACE.format(this::class.java.simpleName)
-        )
+internal interface BaseTransferWorker {
 
     companion object {
         internal const val PART_RECORD_ID = "PART_RECORD_ID"
@@ -86,91 +51,7 @@ internal abstract class BaseTransferWorker(
         internal const val MULTIPART_UPLOAD: String = "MULTIPART_UPLOAD"
     }
 
-    override suspend fun doWork(): Result {
-        // Foreground task is disabled until the foreground notification behavior and the recent customer feedback,
-        // it will be enabled in future based on the customer request.
-        val isForegroundTask: Boolean = (inputData.keyValueMap[RUN_AS_FOREGROUND_TASK] ?: false) as Boolean
-        if (isForegroundTask) {
-            setForegroundAsync(getForegroundInfo())
-        }
-        val result = runCatching {
-            val transferRecordId =
-                inputData.keyValueMap[PART_RECORD_ID] as? Int ?: inputData.keyValueMap[TRANSFER_RECORD_ID] as Int
-            outputData = workDataOf(OUTPUT_TRANSFER_RECORD_ID to inputData.keyValueMap[TRANSFER_RECORD_ID] as Int)
-            transferDB.getTransferRecordById(transferRecordId)?.let { tr ->
-                transferRecord = tr
-                performWork()
-            } ?: return run {
-                Result.failure(outputData)
-            }
-        }
-
-        return when {
-            result.isSuccess -> {
-                result.getOrThrow()
-            }
-            else -> {
-                val ex = result.exceptionOrNull()
-                if (currentCoroutineContext().isActive) {
-                    logger.error("${this.javaClass.simpleName} failed with exception: ${Log.getStackTraceString(ex)}")
-                }
-                if (!currentCoroutineContext().isActive && isRetryableError(ex)) {
-                    Result.retry()
-                } else {
-                    transferStatusUpdater.updateOnError(transferRecord.id, Exception(ex))
-                    transferStatusUpdater.updateTransferState(
-                        transferRecord.id,
-                        TransferState.FAILED
-                    )
-                    Result.failure(outputData)
-                }
-            }
-        }
-    }
-
-    abstract suspend fun performWork(): Result
-
-    internal open var maxRetryCount = 0
-
-    override suspend fun getForegroundInfo(): ForegroundInfo {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            createChannel()
-        }
-        val appIcon = R.drawable.amplify_storage_transfer_notification_icon
-        return ForegroundInfo(
-            1,
-            NotificationCompat.Builder(
-                applicationContext,
-                applicationContext.getString(R.string.amplify_storage_notification_channel_id)
-            )
-                .setSmallIcon(appIcon)
-                .setContentTitle(applicationContext.getString(R.string.amplify_storage_notification_title))
-                .build()
-        )
-    }
-
-    private fun isRetryableError(e: Throwable?): Boolean {
-        return !isNetworkAvailable(applicationContext) ||
-            runAttemptCount < maxRetryCount ||
-            e is CancellationException ||
-            // SocketException is thrown when download is terminated due to network disconnection.
-            e is SocketException
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun createChannel() {
-        val notificationManager =
-            applicationContext.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        notificationManager.createNotificationChannel(
-            NotificationChannel(
-                applicationContext.getString(R.string.amplify_storage_notification_channel_id),
-                applicationContext.getString(R.string.amplify_storage_notification_channel_name),
-                NotificationManager.IMPORTANCE_DEFAULT
-            )
-        )
-    }
-
-    private fun isNetworkAvailable(context: Context): Boolean {
+    fun isNetworkAvailable(context: Context): Boolean {
         val connectivityManager =
             context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
@@ -198,7 +79,7 @@ internal abstract class BaseTransferWorker(
         return false
     }
 
-    internal fun createPutObjectRequest(
+    fun createPutObjectRequest(
         transferRecord: TransferRecord,
         progressListener: ProgressListener?
     ): PutObjectRequest {
